Fixed warning messages
Removed some warning messages from the skeleton code (casts mostly)

+++ wxRaytracer/raytracer/BRDFs/Lambertian.cpp
@@ -52,7 +52,7 @@ Lambertian::operator= (const Lambertian& rhs) {
 
 RGBColor
 Lambertian::f(const ShadeRec& sr, const Vector3D& wo, const Vector3D& wi) const {
-	return (kd * cd * invPI);
+	return (kd * cd * static_cast<const float>(invPI));
 }
