+++ wxRaytracer/raytracer/Materials/Matte.cpp
@@ -88,7 +88,7 @@ Matte::shade(ShadeRec& sr) {
 	
 	for (int j = 0; j < num_lights; j++) {
 		Vector3D wi = sr.w.lights[j]->get_direction(sr);    
-		float ndotwi = sr.normal * wi;
+		float ndotwi = static_cast<float>(sr.normal * wi);
 	
 		if (ndotwi > 0.0) 
 			L += diffuse_brdf->f(sr, wo, wi) * sr.w.lights[j]->L(sr) * ndotwi;
